+++ Editor/BuildConstants.cs
@@ -1,0 +1,19 @@
+namespace BuildScripts.Editor
+{
+    public static class BuildConstants
+    {
+        // Paths
+        public const string BUILD_LOGS_PATH = "../Build/Logs";
+        public const string BUILD_CLIENT_PATH = "../Build/Client";
+
+        // Log files
+        public const string LOG_FILE_PREFIX = "Build-Client-Report";
+        public const string LOG_FILE_EXTENSION = ".log";
+
+        // GUI
+        public const int REMOVE_BUTTON_WIDTH = 70;
+
+        // iOS defaults
+        public const string DEFAULT_IOS_TARGET_VERSION = "13.0";
+    }
+}